requirements.txt
langchain==1.0.8
langchain-core==1.1.0
langchain-community==0.4.1
langchain-cohere==0.5.0
langchain-huggingface==1.0.1
langchain-text-splitters==1.0.0

langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-sdk==0.2.9
langgraph-prebuilt==1.0.5

cohere==5.20.0
chromadb==1.3.5
faiss-cpu==1.13.0

huggingface-hub==0.36.0
tiktoken==0.12.0

python-dotenv==1.2.1
wikipedia==1.4.0
youtube-transcript-api==1.2.3
